[Fix] HCA remove wav error

+++ utils/exporter/converter.go
@@ -40,8 +40,10 @@ func ConvertWavToFLAC(wavFile string, flacFile string, deleteOriginal bool, ffmp
 		return fmt.Errorf("failed to convert WAV to FLAC: %w", err)
 	}
 	if deleteOriginal {
-		if err := os.Remove(wavFile); err != nil {
-			return fmt.Errorf("failed to delete original WAV file: %w", err)
+		if _, err := os.Stat(wavFile); err == nil {
+			if err := os.Remove(wavFile); err != nil {
+				return fmt.Errorf("failed to delete original WAV file: %w", err)
+			}
 		}
 	}
 	return nil
@@ -55,8 +57,10 @@ func ConvertWavToMP3(wavFile string, mp3File string, deleteOriginal bool, ffmpeg
 		return fmt.Errorf("failed to convert WAV to MP3: %w", err)
 	}
 	if deleteOriginal {
-		if err := os.Remove(wavFile); err != nil {
-			return fmt.Errorf("failed to delete original WAV file: %w", err)
+		if _, err := os.Stat(wavFile); err == nil {
+			if err := os.Remove(wavFile); err != nil {
+				return fmt.Errorf("failed to delete original WAV file: %w", err)
+			}
 		}
 	}
 	return nil
